chore: remove reputation/threat-intel enrichment subsystem (OTX + rDNS)
Slim the post-run report pipeline to build/diff/render only. Detect
reports no longer perform any outbound threat-intel lookups.

Removed:
- internal/reputation/* — the pluggable Enricher interface, registry,
  env-driven loader, and the AlienVault OTX backend.
- internal/report/enrich/* — the Source enrichment pipeline (zero
  importers; was future-wiring).
- cmd/coldstep-report enrich.go / reputation.go + the otx-enrich and
  rdns-enrich subcommands; writeModelMap (its only callers) and its test.
- Report model otx / rdns slots and ClassifiedIndicator.rdns field.
- CI enrichment steps (rDNS + OTX) in coldstep-ci-runner.yml; OTX_API_KEY
  / COLDSTEP_OTX_API_KEY usage.

gatherModelIndicators (general render-ip-summary fallback helper, not
reputation-specific) was relocated from the deleted enrich.go into
render.go. writeModelMapFixture (a general test fixture used by six live
render tests) was retained.

Docs updated in lockstep: README, QUICK_START, SECURITY, CLAUDE.md,
scripts/coldstep_detect_report/*, CHANGELOG (Removed). go mod tidy: no
orphaned dependencies. Built clean on Linux (WSL): build, full unit
suite, vet, gofmt, encoding scan all green.
