+++ QUICK_START.md
@@ -226,10 +226,6 @@ Start with default **detect**, then set **`detect-profile: enhanced`** when you 
 
 Consumers copying **`QUICK_START`** alone only need the default digest + JSONL unless they opt into maintainer workflows.
 
-### Optional: OTX enrichment (detect reports)
-
-Set repo/org secret **`OTX_API_KEY`** for AlienVault OTX. Enrichment reads indicators from the active report model (full **`report-model`** or **`ip_classification`** rows on the dev pipeline). No secret → skipped, job still succeeds. Details: **`scripts/coldstep_detect_report/README.md`**.
-
 ---
 
 ## Advanced (optional): previous-run drift diff
