+++ cmd/coldstep-report/model_json.go
@@ -5,9 +5,6 @@ import (
 	"fmt"
 	"io"
 	"os"
-
-	"github.com/coldstep-io/coldstep/internal/atomicwrite"
-	"github.com/coldstep-io/coldstep/internal/report/model"
 )
 
 // Loose upper bound so a hostile or corrupted artifact cannot exhaust memory in-process.
@@ -34,18 +31,4 @@ func readModelMap(path string) (map[string]any, error) {
 		m = map[string]any{}
 	}
 	return m, nil
-}
-
-// writeModelMap persists an enrichment-modified report model using the same
-// canonical encoder as `build-model`: 2-space indent, no HTML escaping, no
-// trailing newline. Plain json.Marshal would re-introduce HTML escaping on
-// `<`/`>`/`&` (silently rewriting PTR records or suspicious domain names) and
-// drop the indentation, diverging from the post-build-model snapshot so
-// downstream attestation hashes stop matching.
-func writeModelMap(path string, m map[string]any) error {
-	raw, err := model.MarshalCanonicalValue(m)
-	if err != nil {
-		return err
-	}
-	return atomicwrite.Bytes(path, raw, 0o644)
 }
