--- internal/reputation/loader/loader_test.go
@@ -1,1 +1,0 @@
-package loader
